+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 - [Branches] Show current branch.
 - [Branches] Switch branch.
 - [Branches] Create and checkout new branch.
+- [Branches] Delete branch.